typescript beer.ts
